Improve test files structure
